Formatting
I consolidated all of my current functions into a single file SHRED_Data. I also created a Fin class to store all of the data that is imported for a fin. I also cleaned up by renaming some of the Notebook files. ScratchSpace is purely a file for testing new functionalities before they are added to the main functions. The other 2 notebooks are examples of how to create an STL file. In the future I want to format them with more text so that they explain what is happening more.